requirements.txt
Flask
requests
opencensus-ext-azure
azure-identity
azure-keyvault-secrets
gunicorn
